fix(dev-core): consolidate workspace-helpers.ts — close traversal-guard gap
- Delete skills/shared/adapters/workspace-helpers.ts (duplicate)
- Redirect imports to cli/lib/workspace-store.ts and github-discovery.ts
- Add traversal guard in detectLocalPath now covers all call sites
- Add lefthook pre-commit grep guard to prevent future drift
- Update types: add type, fieldIds, roadmapProjectId to workspace-store

Closes #101
