+++ plugins/dev-core/skills/init/__tests__/project.test.ts
@@ -30,7 +30,7 @@ vi.mock('../../shared/queries', () => ({
   UPDATE_FIELD_OPTIONS_MUTATION: 'UPDATE_FIELD_OPTIONS_MUTATION',
 }))
 
-vi.mock('../../shared/adapters/workspace-helpers', () => ({
+vi.mock('../../../cli/lib/workspace-store', () => ({
   readWorkspace: vi.fn(() => ({ projects: [] })),
   writeWorkspace: vi.fn(),
   getWorkspacePath: () => '/tmp/test-workspace.json',
