+++ plugins/dev-core/skills/init/lib/project.ts
@@ -3,6 +3,7 @@
  * Supports --type technical|company to configure per-project field slots.
  */
 
+import { readWorkspace, writeWorkspace } from '../../../cli/lib/workspace-store'
 import {
   DEFAULT_PRIORITY_OPTIONS,
   DEFAULT_SIZE_OPTIONS,
@@ -10,7 +11,6 @@ import {
 } from '../../shared/adapters/config-helpers'
 import type { ParsedField } from '../../shared/adapters/github-adapter'
 import { ghGraphQL, linkProjectToRepo, parseProjectFields, run } from '../../shared/adapters/github-adapter'
-import { readWorkspace, writeWorkspace } from '../../shared/adapters/workspace-helpers'
 import type { ProjectFieldIds } from '../../shared/domain/types'
 import type { ProjectType, WorkspaceProject } from '../../shared/ports/workspace'
 import { PROJECT_WORKFLOWS_QUERY, UPDATE_FIELD_OPTIONS_MUTATION } from '../../shared/queries'
