--- plugins/dev-core/skills/shared/adapters/workspace-helpers.ts
@@ -1,85 +1,0 @@
-/**
- * Pure workspace helper functions extracted from the legacy workspace.ts shim.
- * These are used by GitWorkspaceAdapter and re-exported from workspace.ts for backward compat.
- */
-import { existsSync, mkdirSync, readFileSync, writeFileSync } from 'node:fs'
-import type { Workspace, WorkspaceProject } from '../ports/workspace'
-
-export function getWorkspacePath(): string {
-  const home = process.env.HOME
-  if (!home) throw new Error('HOME environment variable is not set')
-  const vault = `${home}/.roxabi-vault`
-  if (existsSync(vault)) return `${vault}/workspace.json`
-  return `${home}/.config/roxabi/workspace.json`
-}
-
-export function readWorkspace(): Workspace {
-  const p = getWorkspacePath()
-  if (!existsSync(p)) return { projects: [] }
-  try {
-    return JSON.parse(readFileSync(p, 'utf8')) as Workspace
-  } catch {
-    return { projects: [] }
-  }
-}
-
-export function writeWorkspace(ws: Workspace): void {
-  const p = getWorkspacePath()
-  const dir = p.substring(0, p.lastIndexOf('/'))
-  if (!existsSync(dir)) {
-    mkdirSync(dir, { recursive: true, mode: 0o700 })
-  }
-  writeFileSync(p, `${JSON.stringify(ws, null, 2)}\n`, { mode: 0o600 })
-}
-
-/**
- * Discover GitHub Projects linked to a repo via GraphQL.
- * Returns array of WorkspaceProject (may be 0, 1, or multiple).
- */
-export async function discoverProject(repo: string): Promise<WorkspaceProject[]> {
-  const [owner, name] = repo.split('/')
-  if (!owner || !name) throw new Error(`Invalid repo format: '${repo}'. Use 'owner/name'.`)
-
-  const query = `query($owner: String!, $name: String!) {
-    repository(owner: $owner, name: $name) {
-      projectsV2(first: 10) { nodes { id title } }
-    }
-  }`
-
-  const token =
-    process.env.GITHUB_TOKEN ||
-    (() => {
-      const proc = Bun.spawnSync(['gh', 'auth', 'token'], { stdout: 'pipe', stderr: 'pipe' })
-      return new TextDecoder().decode(proc.stdout).trim()
-    })()
-  if (!token) throw new Error('Not authenticated. Run: gh auth login or set GITHUB_TOKEN env var')
-
-  const res = await fetch('https://api.github.com/graphql', {
-    method: 'POST',
-    headers: { Authorization: `Bearer ${token}`, 'Content-Type': 'application/json' },
-    body: JSON.stringify({ query, variables: { owner, name } }),
-  })
-  if (!res.ok) throw new Error(`GitHub API error: ${res.status} ${res.statusText}`)
-
-  const json = (await res.json()) as {
-    data: { repository: { projectsV2: { nodes: { id: string; title: string }[] } } }
-  }
-  const localPath = detectLocalPath(name)
-  return (json.data?.repository?.projectsV2?.nodes ?? []).map((n) => ({
-    repo,
-    projectId: n.id,
-    label: n.title,
-    ...(localPath ? { localPath } : {}),
-  }))
-}
-
-/** Try to find the local clone of a repo by scanning common directories. */
-function detectLocalPath(repoName: string): string | undefined {
-  const home = process.env.HOME
-  if (!home) return undefined
-  const candidates = [`${home}/projects/${repoName}`, `${home}/${repoName}`, `${home}/src/${repoName}`]
-  for (const dir of candidates) {
-    if (existsSync(`${dir}/.git`)) return dir
-  }
-  return undefined
-}
